+++ packages/k-skill-proxy/package.json
@@ -9,7 +9,7 @@
     "node": ">=20"
   },
   "scripts": {
-    "lint": "node --check src/airkorea.js && node --check src/assembly.js && node --check src/building-register.js && node --check src/ev-charger.js && node --check src/hrfco.js && node --check src/keris-academic.js && node --check src/korean-holiday.js && node --check src/korean-law.js && node --check src/kopis.js && node --check src/kr-whois.js && node --check src/krx-stock.js && node --check src/kstartup.js && node --check src/kakao-map.js && node --check src/lh-notice.js && node --check src/mfds.js && node --check src/molit.js && node --check src/naver-news.js && node --check src/nhis-care.js && node --check src/g2b-sanction.js && node --check src/g2b-order-plan.js && node --check src/fsc-corp.js && node --check src/national-pension.js && node --check src/naver-shopping.js && node --check src/nts-business.js && node --check src/parking-lots.js && node --check src/region-lookup.js && node --check src/server.js && node --check test/airkorea.test.js && node --check test/building-register.test.js && node --check test/ev-charger.test.js && node --check test/hrfco.test.js && node --check test/keris-academic.test.js && node --check test/korean-law.test.js && node --check test/lh-notice.test.js && node --check test/molit.test.js && node --check test/naver-news.test.js && node --check test/naver-shopping.test.js && node --check test/region-lookup.test.js && node --check test/server.test.js && node --check test/g2b-order-plan.test.js",
+    "lint": "node --check src/airkorea.js && node --check src/assembly.js && node --check src/building-register.js && node --check src/ev-charger.js && node --check src/hrfco.js && node --check src/keris-academic-xml.js && node --check src/keris-academic.js && node --check src/korean-holiday.js && node --check src/korean-law.js && node --check src/kopis.js && node --check src/kr-whois.js && node --check src/krx-stock.js && node --check src/kstartup.js && node --check src/kakao-map.js && node --check src/lh-notice.js && node --check src/mfds.js && node --check src/molit.js && node --check src/naver-news.js && node --check src/nhis-care.js && node --check src/g2b-sanction.js && node --check src/g2b-order-plan.js && node --check src/fsc-corp.js && node --check src/national-pension.js && node --check src/naver-shopping.js && node --check src/nts-business.js && node --check src/parking-lots.js && node --check src/region-lookup.js && node --check src/server.js && node --check test/airkorea.test.js && node --check test/building-register.test.js && node --check test/ev-charger.test.js && node --check test/hrfco.test.js && node --check test/keris-academic.test.js && node --check test/korean-law.test.js && node --check test/lh-notice.test.js && node --check test/molit.test.js && node --check test/naver-news.test.js && node --check test/naver-shopping.test.js && node --check test/region-lookup.test.js && node --check test/server.test.js && node --check test/g2b-order-plan.test.js",
     "test": "node --test"
   },
   "dependencies": {
