+++ packages/k-skill-proxy/src/keris-academic.js
@@ -7,6 +7,7 @@ const RESOURCE_TYPE_MAP = Object.freeze({
   B: ["U"]
 });
 const SEARCH_FIELDS = ["keyword", "title", "author", "subject", "publisher"];
+const { parseRissXml } = require("./keris-academic-xml");
 
 function trimOrNull(value) {
   if (value === undefined || value === null) return null;
@@ -74,163 +75,6 @@ function normalizeKerisAcademicQuery(query = {}) {
     pageSize,
     rsnum: ((page - 1) * pageSize) + 1
   };
-}
-
-function decodeXmlEntities(value) {
-  return value.replace(/&(#x[0-9a-f]+|#\d+|amp|lt|gt|quot|apos);/gi, (match, entity) => {
-    const lower = entity.toLowerCase();
-    if (lower === "amp") return "&";
-    if (lower === "lt") return "<";
-    if (lower === "gt") return ">";
-    if (lower === "quot") return '"';
-    if (lower === "apos") return "'";
-    const codePoint = lower.startsWith("#x")
-      ? Number.parseInt(lower.slice(2), 16)
-      : Number.parseInt(lower.slice(1), 10);
-    return Number.isFinite(codePoint) ? String.fromCodePoint(codePoint) : match;
-  });
-}
-
-function parseXmlDocument(xml) {
-  const source = trimOrNull(xml);
-  if (!source || !source.startsWith("<")) throw new Error("RISS upstream did not return XML.");
-  if (/<!DOCTYPE|<!ENTITY/i.test(source)) throw new Error("RISS upstream XML contains unsupported declarations.");
-  const document = { name: "#document", children: [], text: "" };
-  const stack = [document];
-  const tokens = source.match(/<!--[^]*?-->|<!\[CDATA\[[^]*?\]\]>|<\?[^]*?\?>|<[^>]+>|[^<]+/g);
-  if (!tokens || tokens.join("") !== source) throw new Error("RISS upstream returned malformed XML.");
-  for (const token of tokens) {
-    if (token.startsWith("<!--") || token.startsWith("<?")) continue;
-    if (token.startsWith("<![CDATA[")) {
-      stack[stack.length - 1].text += token.slice(9, -3);
-    } else if (!token.startsWith("<")) {
-      stack[stack.length - 1].text += decodeXmlEntities(token);
-    } else if (token.startsWith("</")) {
-      const name = token.slice(2, -1).trim();
-      if (stack.length === 1 || stack[stack.length - 1].name !== name) throw new Error("RISS upstream returned malformed XML.");
-      stack.pop();
-    } else {
-      if (token.startsWith("<!")) throw new Error("RISS upstream XML contains unsupported markup.");
-      const selfClosing = /\/\s*>$/.test(token);
-      const match = token.match(/^<([A-Za-z_][\w:.-]*)(?:\s[^>]*)?\/?\s*>$/);
-      if (!match) throw new Error("RISS upstream returned malformed XML.");
-      const node = { name: match[1], children: [], text: "" };
-      stack[stack.length - 1].children.push(node);
-      if (!selfClosing) stack.push(node);
-    }
-  }
-  if (stack.length !== 1 || document.children.length !== 1) throw new Error("RISS upstream returned malformed XML.");
-  return document.children[0];
-}
-
-function child(node, name) {
-  return node.children.find((entry) => entry.name.toLowerCase() === name.toLowerCase()) || null;
-}
-
-function children(node, name) {
-  return node.children.filter((entry) => entry.name.toLowerCase() === name.toLowerCase());
-}
-
-function nodeText(node) {
-  if (!node) return null;
-  const nested = node.children.map(nodeText).filter(Boolean).join(" ");
-  return trimOrNull(`${node.text}${nested ? ` ${nested}` : ""}`);
-}
-
-function metadataValues(node) {
-  const values = new Map();
-  for (const entry of node.children) {
-    const key = entry.name.replace(/^riss\./i, "").toLowerCase();
-    const value = nodeText(entry);
-    if (value === null) continue;
-    if (!values.has(key)) values.set(key, []);
-    values.get(key).push(value);
-  }
-  return values;
-}
-
-function first(values, key) {
-  return values.get(key)?.[0] ?? null;
-}
-
-function splitAuthors(value) {
-  if (!value) return [];
-  const explicit = value.split(/\s*[;,|]\s*/).filter(Boolean);
-  if (explicit.length > 1) return explicit;
-  const koreanNames = value.split(/\s+/).filter(Boolean);
-  if (koreanNames.length > 1 && koreanNames.every((name) => /^[가-힣]{2,4}$/.test(name))) return koreanNames;
-  return [value];
-}
-
-function normalizeMetadata(node) {
-  const values = metadataValues(node);
-  const image = first(values, "image")?.toUpperCase() ?? null;
-  const charge = first(values, "charge");
-  const fullTextAvailable = image === "Y" ? true : image === "N" ? false : null;
-  let fullTextAccess = "unknown";
-  if (fullTextAvailable === false) fullTextAccess = "none";
-  if (fullTextAvailable === true && charge === "1") fullTextAccess = "free";
-  if (fullTextAvailable === true && charge === "0") fullTextAccess = "paid_or_restricted";
-  if (fullTextAvailable === true && charge === null) fullTextAccess = "available";
-  return {
-    resource_type: first(values, "type"),
-    title: first(values, "title"),
-    authors: splitAuthors(first(values, "author")),
-    publisher: first(values, "publisher"),
-    year: first(values, "pubdate"),
-    publication: first(values, "stitle"),
-    material_type: first(values, "mtype"),
-    link: first(values, "url"),
-    full_text_available: fullTextAvailable,
-    full_text_access: fullTextAccess,
-    holdings: values.get("holdings") || [],
-  };
-}
-
-function parseRissXml(xml) {
-  let root;
-  try {
-    root = parseXmlDocument(xml);
-  } catch (error) {
-    error.code = "upstream_invalid_response";
-    throw error;
-  }
-  if (root.name.toLowerCase() !== "record") {
-    const error = new Error("RISS upstream returned an invalid response envelope.");
-    error.code = "upstream_invalid_response";
-    throw error;
-  }
-  const head = child(root, "head");
-  if (!head) {
-    const error = new Error("RISS upstream response is missing head metadata.");
-    error.code = "upstream_invalid_response";
-    throw error;
-  }
-  const errorCode = nodeText(child(head, "Error"));
-  if (errorCode === null) {
-    const error = new Error("RISS upstream response is missing status metadata.");
-    error.code = "upstream_invalid_response";
-    throw error;
-  }
-  const errorMessage = nodeText(child(head, "ErrorMessage")) || "Unknown RISS error";
-  if (errorCode !== null && !["0", "000"].includes(errorCode)) {
-    const error = new Error(errorMessage);
-    error.code = ["004", "4"].includes(errorCode) || /인증|AUTH|KEY/i.test(errorMessage)
-      ? "upstream_forbidden"
-      : /쿼터|호출량|초과|QUOTA|LIMIT/i.test(errorMessage)
-        ? "upstream_quota_exceeded"
-        : "upstream_error";
-    error.upstreamCode = errorCode;
-    throw error;
-  }
-  const items = children(root, "metadata").map(normalizeMetadata);
-  const totalCount = Number.parseInt(nodeText(child(head, "totalcount")) ?? String(items.length), 10);
-  if (!Number.isFinite(totalCount) || totalCount < 0) {
-    const error = new Error("RISS upstream returned invalid totalcount metadata.");
-    error.code = "upstream_invalid_response";
-    throw error;
-  }
-  return { totalCount, items };
 }
 
 function errorResult(error, message) {
